refactor: change pages file structure
