+++ frontend/src/App.tsx
@@ -4,7 +4,7 @@ import {CourseDto} from "./types/courseTypes.ts";
 import {Route, Routes, useNavigate} from "react-router-dom";
 import CourseDetailsPage from "./pages/CourseDetails/CourseDetailsPage.tsx";
 import CourseCreator from "./pages/CourseCreator.tsx";
-import Dashboard from "./pages/Dashboard/Dashboard.tsx";
+import Dashboard from "./pages/Dashboard.tsx";
 import LessonOverview from "./pages/CourseDetails/Lesson/LessonOverview.tsx";
 import AssignmentOverview from "./pages/CourseDetails/Assignment/AssignmentOverview.tsx";
 import LessonPage from "./pages/CourseDetails/Lesson/LessonPage.tsx";
@@ -24,7 +24,7 @@ import UserAccountPage from "./pages/UserAccountPage.tsx";
 import {createTheme, ThemeProvider} from "@mui/material";
 import {themeOptions} from "./styles/themeOptions.ts";
 import ParticipantOverview from "./pages/CourseDetails/Participant/ParticipantOverview.tsx";
-import BrowsePage from "./pages/BrowsePage/BrowsePage.tsx";
+import BrowsePage from "./pages/BrowsePage.tsx";
 import axiosInstance from "./api/axiosInstance.ts";
 import axios from "axios";
 import {useDataArray} from "./hooks/useDataArray.ts";
